@@ -1,9 +1,7 @@
 {
   "name": "@skyux/grids",
-  "version": "3.8.0",
+  "version": "4.0.0",
   "description": "SKY UX Grids",
-  "main": "bundles/bundle.umd.js",
-  "module": "index.ts",
   "scripts": {
     "build": "skyux build-public-library",
     "skyux": "skyux",
@@ -23,61 +21,57 @@
   },
   "homepage": "https://github.com/blackbaud/skyux-grids#readme",
   "peerDependencies": {
-    "@angular/common": ">=4.3.6",
-    "@angular/core": ">=4.3.6",
-    "@angular/forms": ">=4.3.6",
-    "@skyux/core": "^3.5.1",
-    "@skyux/forms": "^3.0.0",
-    "@skyux/i18n": "^3.5.0",
-    "@skyux/indicators": "^3.0.0",
-    "@skyux/layout": "^3.5.0",
-    "@skyux/list-builder-common": "^3.0.0",
-    "@skyux/popovers": "^3.4.0"
+    "@angular/common": "^9.1.7",
+    "@angular/core": "^9.1.7",
+    "@angular/forms": "^9.1.7",
+    "@skyux/core": "^4.0.0",
+    "@skyux/forms": "^4.0.0",
+    "@skyux/i18n": "^4.0.0",
+    "@skyux/indicators": "^4.0.0",
+    "@skyux/layout": "^4.0.0",
+    "@skyux/list-builder-common": "^4.0.0",
+    "@skyux/popovers": "^4.0.0"
   },
   "dependencies": {
     "dragula": "3.7.2",
     "ng2-dragula": "1.5.0"
   },
   "devDependencies": {
-    "@angular/animations": "7.2.16",
-    "@angular/common": "7.2.16",
-    "@angular/compiler": "7.2.16",
-    "@angular/compiler-cli": "7.2.16",
-    "@angular/core": "7.2.16",
-    "@angular/forms": "7.2.16",
-    "@angular/http": "7.2.16",
-    "@angular/platform-browser": "7.2.16",
-    "@angular/platform-browser-dynamic": "7.2.16",
-    "@angular/router": "7.2.16",
+    "@angular/animations": "9.1.7",
+    "@angular/common": "9.1.7",
+    "@angular/compiler": "9.1.7",
+    "@angular/compiler-cli": "9.1.7",
+    "@angular/core": "9.1.7",
+    "@angular/forms": "9.1.7",
+    "@angular/platform-browser": "9.1.7",
+    "@angular/platform-browser-dynamic": "9.1.7",
+    "@angular/router": "9.1.7",
     "@blackbaud/auth-client": "2.24.0",
-    "@skyux-sdk/builder": "3.17.0",
-    "@skyux-sdk/builder-plugin-skyux": "1.3.3",
-    "@skyux-sdk/e2e": "3.2.0",
-    "@skyux-sdk/testing": "3.3.0",
-    "@skyux/animations": "3.0.0",
-    "@skyux/assets": "3.1.0",
-    "@skyux/config": "3.9.0",
-    "@skyux/core": "3.14.2",
-    "@skyux/forms": "3.6.3",
-    "@skyux/grids": "3.6.0",
-    "@skyux/http": "3.9.0",
-    "@skyux/i18n": "3.11.0",
-    "@skyux/indicators": "3.2.1",
-    "@skyux/layout": "3.9.0",
-    "@skyux/list-builder-common": "3.0.0",
-    "@skyux/modals": "3.2.0",
-    "@skyux/omnibar-interop": "3.1.0",
-    "@skyux/popovers": "3.5.3",
-    "@skyux/router": "3.1.1",
-    "@skyux/theme": "3.15.0",
-    "@types/core-js": "2.5.3",
-    "codelyzer": "4.5.0",
-    "core-js": "2.6.11",
+    "@skyux-sdk/builder": "4.0.0-rc.9",
+    "@skyux-sdk/builder-plugin-skyux": "4.0.0-rc.0",
+    "@skyux-sdk/e2e": "4.0.0-rc.1",
+    "@skyux-sdk/testing": "4.0.0-rc.3",
+    "@skyux/animations": "4.0.0",
+    "@skyux/assets": "4.0.0",
+    "@skyux/config": "4.0.0",
+    "@skyux/core": "4.0.0",
+    "@skyux/forms": "4.0.0",
+    "@skyux/http": "4.0.0",
+    "@skyux/i18n": "4.0.0",
+    "@skyux/indicators": "4.0.0",
+    "@skyux/layout": "4.0.0",
+    "@skyux/list-builder-common": "4.0.0",
+    "@skyux/modals": "4.0.0",
+    "@skyux/omnibar-interop": "4.0.0",
+    "@skyux/popovers": "4.0.0",
+    "@skyux/router": "4.0.0",
+    "@skyux/theme": "4.0.0",
+    "codelyzer": "5.2.2",
+    "moment": "2.25.3",
     "rxjs": "6.5.5",
-    "rxjs-compat": "6.5.5",
-    "ts-node": "8.9.1",
-    "tslint": "5.20.1",
-    "typescript": "3.2.4",
-    "zone.js": "0.8.28"
+    "ts-node": "8.3.0",
+    "tslint": "6.1.2",
+    "typescript": "3.8.3",
+    "zone.js": "0.10.3"
   }
 }
