@@ -3,10 +3,6 @@ import {
 } from '@angular/core';
 
 import {
-  SkyPopoverComponent
-} from '@skyux/popovers';
-
-import {
   SkyGridColumnAlignment
 } from './types/grid-column-alignment';
 
@@ -15,7 +11,7 @@ export class SkyGridColumnModel {
   public id: string;
   public field: string;
   public heading: string;
-  public inlineHelpPopover: SkyPopoverComponent;
+  public inlineHelpPopover: any;
   public type: string;
   public width: number;
   public hidden: boolean;
